Full code. Verify correctness. Write codes/scripts for tests.

+++ src/Global.jl
@@ -34,6 +34,7 @@ function sample(rng::AbstractRNG, ud::UniformDist)
   return lb + Base.rand(rng) * fact
 end
 
+#Given car physical state, track the nearest leading and following vehicle in each lane.
 typealias NbCache Dict{CarPhysicalState, Array{SVector{2,CarPhysicalState},1}}
 
 #State of world. Should work for upper level as well.
@@ -44,7 +45,7 @@ type GlobalStateL1
   _neighbor_cache::NbCache
 end
 GlobalStateL1(ego::CarPhysicalState, neighborhood::Array{Array{CarLocalIS,1},1}) = GlobalStateL1(0, ego, neighborhood, NbCache())
-GlobalStateL1(term, ego::CarPhysicalState, neighborhood::Array{Array{CarLocalIS,1},1}) = GlobalStateL1(term, ego, neighborhood, NbCache())
+GlobalStateL1(term::Int64, ego::CarPhysicalState, neighborhood::Array{Array{CarLocalIS,1},1}) = GlobalStateL1(term, ego, neighborhood, NbCache())
 ==(s1::GlobalStateL1,s2::GlobalStateL1) = (s1.terminal == s2.terminal) && (s1.ego == s2.ego) && (s1.neighborhood == s2.neighborhood)
 Base.hash(s1::GlobalStateL1, h::UInt64 = zero(UInt64)) = hash(s1.terminal, hash(s1.ego, hash(s1.neighborhood, h)))
 Base.copy(s1::GlobalStateL1) = GlobalStateL1(s1.terminal, s1.ego, s1.neighborhood)
